Added player parsing (Save part)

+++ Game/Player/Player.hpp
@@ -37,9 +37,9 @@ namespace Game
             * @param position A const reference to a std::pair of float
             * @param brokenWalls A const reference to a size_t
             * @param kills A const reference to a size_t
-            * @param powerUps A const reference to a std::array of 5 size_t
+            * @param powerUps A const reference to a std::array of 5 int
             */
-            Player(const std::string &name, const std::pair<float, float> &positions, const size_t &kills, const size_t &brokenWalls, const std::array<size_t, 5> &powerUps);
+            Player(const std::string &name, const std::pair<float, float> &positions, const size_t &kills, const size_t &brokenWalls, const std::array<int, 5> &powerUps);
 
             /**
             * @brief Destroy the Player object
@@ -77,9 +77,9 @@ namespace Game
             /**
              * @brief Get PowerUps gotten by the Player
              *
-             * @return An Array of 5 (size_t)
+             * @return An Array of 5 (int)
              */
-            std::array<size_t, 5> getPowerUps(void) const;
+            std::array<int, 5> getPowerUps(void) const;
 
             /**
              * @brief Get the player ID
@@ -119,9 +119,9 @@ namespace Game
             /**
              * @brief Set the powerUps of the player
              *
-             * @param powerUps A const reference to an array of 5 size_t
+             * @param powerUps A const reference to an array of 5 int
              */
-            void setPowerUps(const std::array<size_t, 5> &powerUps);
+            void setPowerUps(const std::array<int, 5> &powerUps);
 
             /**
              * @brief Set the ID of the player
@@ -133,7 +133,7 @@ namespace Game
             std::pair<float, float> _positions; // A pair of float for player positions (x, y)
             size_t _brokenWalls; // Amount of broken walls
             size_t _kills; // Amount of kills
-            std::array<size_t, 5> _powerUps; // Amount of powerUps
+            std::array<int, 5> _powerUps; // Amount of powerUps
             int _ID; // ID of the player
     };
