+++ Game/Player/Player.inl
@@ -25,7 +25,7 @@ inline size_t Game::Player::getKills(void) const
     return _kills;
 }
 
-inline std::array<size_t, 5> Game::Player::getPowerUps(void) const
+inline std::array<int, 5> Game::Player::getPowerUps(void) const
 {
     return _powerUps;
 }
@@ -55,7 +55,7 @@ inline void Game::Player::setKills(const size_t &kills)
     _kills = kills;
 }
 
-inline void Game::Player::setPowerUps(const std::array<size_t, 5> &powerUps)
+inline void Game::Player::setPowerUps(const std::array<int, 5> &powerUps)
 {
     _powerUps = powerUps;
 }
